stock_scanner sentinel: fix menu display
when the number of elements to display is long, the
'Selected' display is written on top of the last element.
This makes selecting the last element of the string difficult
because the label can be completely hidden from view.

This fixes the issue by reserving the last line for the Selection
display.